backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain
langchain-community
dateparser
google-api-python-client
google-auth
google-auth-oauthlib
